requirements.txt
Flask-RQ==0.2
Flask-Redis==0.0.5
Flask-Script==2.0.5
Jinja2==2.7.3
MarkupSafe==0.23
Unidecode==0.04.16
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
python-slugify==0.0.7
redis==2.10.1
requests==2.3.0
rom==0.26.5
rq==0.4.6
six==1.7.3
wsgiref==0.1.2
